[#6] refactor: 음원 게시물 관련(레포지토리) 추상화
