+++ module-core/src/main/java/com/soundie/post/repository/PostLikeRepository.java
@@ -1,57 +1,19 @@
 package com.soundie.post.repository;
 
 import com.soundie.post.domain.PostLike;
-import org.springframework.stereotype.Repository;
 
-import java.util.*;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.List;
+import java.util.Optional;
 
-@Repository
-public class PostLikeRepository {
+public interface PostLikeRepository {
 
-    private final Map<Long, PostLike> store = new ConcurrentHashMap<>();
-    private AtomicLong sequence = new AtomicLong(0L);
+    List<PostLike> findPostLikes();
 
-    /*
-     * 좋아요 목록 조회
-     * */
-    public List<PostLike> findPostLikes() {
-        return new ArrayList<>(store.values());
-    }
+    Optional<PostLike> findPostLikeByMemberIdAndPostId(Long memberId, Long postId);
 
-    /*
-     * 회원 Id + 음원 게시물 Id로, 좋아요 조회
-     * */
-    public Optional<PostLike> findPostLikeByMemberIdAndPostId(Long memberId, Long postId) {
-        return findPostLikes().stream()
-                .filter(pl -> pl.getMemberId().equals(memberId) && pl.getPostId().equals(postId) )
-                .findFirst();
-    }
+    Long countPostLikesByPostId(Long postId);
 
-    /*
-     * 좋아요 개수 조회
-     * */
-    public Long countPostLikesByPostId(Long postId){
-        return findPostLikes().stream()
-                .filter(pl -> pl.getPostId().equals(postId))
-                .count();
-    }
+    PostLike save(PostLike postLike);
 
-    /*
-     * 좋아요 저장
-     * */
-    public PostLike save(PostLike postLike) {
-        postLike.setId(sequence.incrementAndGet());
-        store.put(postLike.getId(), postLike);
-
-        return postLike;
-    }
-
-    /*
-    * 좋아요 삭제
-    * */
-    public void delete(PostLike postLike){
-        store.remove(postLike.getId());
-    }
+    void delete(PostLike postLike);
 }
