+++ module-core/src/main/java/com/soundie/post/repository/PostRepository.java
@@ -1,46 +1,15 @@
 package com.soundie.post.repository;
 
 import com.soundie.post.domain.Post;
-import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicLong;
 
-@Repository
-public class PostRepository {
+public interface PostRepository {
 
-    private final Map<Long, Post> store = new ConcurrentHashMap<>();
-    private AtomicLong sequence = new AtomicLong(0L);
+    List<Post> findPosts();
 
-    /*
-    * 음원 게시물 목록 조회
-    * */
-    public List<Post> findPosts() {
-        return new ArrayList<>(store.values());
-    }
+    Optional<Post> findPostById(Long postId);
 
-    /*
-     * 음원 게시물 Id로, 음원 게시물 조회
-     * */
-    public Optional<Post> findPostById(Long postId) {
-        return Optional.ofNullable(store.get(postId));
-    }
-
-    /*
-    * 음원 게시물 저장
-    * */
-    public Post save(Post post){
-        post.setId(sequence.incrementAndGet());
-        store.put(post.getId(), post);
-
-        return post;
-    }
-
-    public void clearStore() {
-        store.clear();
-    }
+    Post save(Post post);
 }
